refactor - move constants to constants.py
